Global commit

+++ app/controllers/dossiers_controller.rb
@@ -2,11 +2,13 @@ class DossiersController < ApplicationController
   before_action :set_carton
 def new
     @dossier = @carton.dossiers.build
+    authorize @carton
   end
 
   def create
     @dossier = Dossier.new(dossier_params)
     @dossier.carton = @carton
+    authorize @carton
     if @dossier.save
       redirect_to carton_path(@carton)
     else
@@ -15,11 +17,13 @@ def new
   end
 
   def edit
+    authorize @carton
     @carton = Carton.find(params[:carton_id])
     @dossier = Dossier.find(params[:id])
   end
 
   def update
+    authorize @carton
     @carton = Carton.find(params[:carton_id])
     @dossier = Dossier.find(params[:id])
     @dossier.update(dossier_params)
@@ -28,10 +32,11 @@ def new
   end
 
     def destroy
+    authorize @carton
     # TO DO : flash notice s'affiche 2 fois
     @carton = Carton.find(params[:carton_id])
     @dossier = Dossier.find(params[:id])
-    title = @dossier.name
+    title = @dossier.sujet
     if @dossier.destroy
       flash[:notice] = "#{title} a été supprimé avec succès."
       redirect_to @carton
@@ -48,6 +53,6 @@ def new
   end
 
   def dossier_params
-    params.require(:dossier).permit(:name, :client)
+    params.require(:dossier).permit(:sujet, :client, :sensible)
   end
 end
